+++ tests/e2e/core-tests/specs/index.js
@@ -16,9 +16,11 @@ const runCheckoutApplyCouponsTest = require( './shopper/front-end-checkout-coupo
 const runCheckoutPageTest = require( './shopper/front-end-checkout.test' );
 const runMyAccountPageTest = require( './shopper/front-end-my-account.test' );
 const runMyAccountPayOrderTest = require( './shopper/front-end-my-account-pay-order.test' );
+const runMyAccountCreateAccountTest = require( './shopper/front-end-my-account-create-account.test' );
 const runSingleProductPageTest = require( './shopper/front-end-single-product.test' );
 const runVariableProductUpdateTest = require( './shopper/front-end-variable-product-updates.test' );
 const runCheckoutCreateAccountTest = require( './shopper/front-end-checkout-create-account.test' );
+const runCheckoutLoginAccountTest = require( './shopper/front-end-checkout-login-account.test' );
 
 // Merchant tests
 const runAddNewShippingZoneTest = require ( './merchant/wp-admin-settings-shipping-zones.test' );
@@ -60,9 +62,11 @@ const runShopperTests = () => {
 	runCheckoutPageTest();
 	runMyAccountPageTest();
 	runMyAccountPayOrderTest();
+	runMyAccountCreateAccountTest();
 	runSingleProductPageTest();
 	runVariableProductUpdateTest();
 	runCheckoutCreateAccountTest();
+	runCheckoutLoginAccountTest();
 };
 
 const runMerchantTests = () => {
@@ -133,4 +137,6 @@ module.exports = {
 	runApiTests,
 	runAnalyticsPageLoadsTest,
 	runCheckoutCreateAccountTest,
+	runCheckoutLoginAccountTest,
+	runMyAccountCreateAccountTest,
 };