+++ tests/e2e/core-tests/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 # 0.1.3
 
+## Added
+
+- Shopper Checkout Login Account
+- Shopper My Account Create Account
+
 ## Fixed
 
 - removed use of ES6 `import`